separate app to another file, in server.js live database connection configuration
